Thu, Jun 20, 2024, 3:21 PM -04:00

+++ readability/readability.c
@@ -32,7 +32,7 @@ int main() {
     float S = (float)sentances / words * 100;
 
     // calcuate index
-    int index = 0.0588 * L - 0.296 * S -15.8;
+    float index = 0.0588 * L - 0.296 * S -15.8;
 
     // print result
     if (index < 1) {
@@ -40,7 +40,7 @@ int main() {
     } else if (index >= 16){
         printf("Grade 16+\n");
     } else {
-        printf("Grade %d\n", index);
+        printf("Grade %f\n", index);
     }
 
     return 0;
